Ubuntu: Bionic: Add a different image for the HP Envy x2
Annoyingly there only 2b (bits) of difference between the two images.

The HP Envy x2 has an inverted SD card-detect GPIO line when compared
to the other devices.  A very small change is required in the DTB in
order for Linux to locate the installed SD card containing the rootfs.

Signed-off-by: AArch64 Laptops <aarch64laptops@gmail.com>